chore: CEXT-1487: refactored code
