+++ src/commands/__fixtures__/sample_mesh_files.json
@@ -1,23 +1,23 @@
 {
-  "meshConfig":{
-     "sources":[
-        {
-           "name":"<json_source_name>",
-           "handler":{
-              "JsonSchema":{
-                 "baseUrl":"<json_source__baseurl>",
-                 "operations":[
-                    {
-                       "type":"Query",
-                       "field":"<query>",
-                       "path":"<query_path>",
-                       "method":"POST",
-                       "requestSchema":"./requestParams.json"
-                    }
-                 ]
+  "meshConfig": {
+    "sources": [
+      {
+        "name": "<json_source_name>",
+        "handler": {
+          "JsonSchema": {
+            "baseUrl": "<json_source__baseurl>",
+            "operations": [
+              {
+                "type": "Query",
+                "field": "<query>",
+                "path": "<query_path>",
+                "method": "POST",
+                "requestSchema": "./requestParams.json"
               }
-           }
+            ]
+          }
         }
-     ]
+      }
+    ]
   }
 }
