+++ src/commands/__fixtures__/sample_mesh_invalid_paths.json
@@ -1,0 +1,23 @@
+{
+    "meshConfig": {
+      "sources": [
+        {
+          "name": "<json_source_name>",
+          "handler": {
+            "JsonSchema": {
+              "baseUrl": "<json_source__baseurl>",
+              "operations": [
+                {
+                  "type": "Query",
+                  "field": "<query>",
+                  "path": "<query_path>",
+                  "method": "POST",
+                  "requestSchema": "./schemaBody.json"
+                }
+              ]
+            }
+          }
+        }
+      ]
+    }
+  }
