+++ src/commands/__fixtures__/sample_mesh_with_files_array.json
@@ -1,0 +1,29 @@
+{
+    "meshConfig": {
+      "sources": [
+        {
+          "name": "<json_source_name>",
+          "handler": {
+            "JsonSchema": {
+              "baseUrl": "<json_source__baseurl>",
+              "operations": [
+                {
+                  "type": "Query",
+                  "field": "<query>",
+                  "path": "<query_path>",
+                  "method": "POST",
+                  "requestSchema": "./requestParams.json"
+                }
+              ]
+            }
+          }
+        }
+      ],
+      "files": [
+        {
+          "path": "./requestParams.json",
+          "content": "{\"type\":\"dummyContent\"}"
+        }
+      ]
+    }
+  }
