+++ src/commands/api-mesh/__tests__/create.test.js
@@ -14,13 +14,7 @@ const mockConsoleCLIInstance = {};
 
 const CreateCommand = require('../create');
 const sampleCreateMeshConfig = require('../../__fixtures__/sample_mesh.json');
-const {
-	initSdk,
-	initRequestId,
-	promptConfirm,
-	getFilesInMeshConfig,
-	importFiles,
-} = require('../../../helpers');
+const { initSdk, initRequestId, promptConfirm, importFiles } = require('../../../helpers');
 const {
 	createMesh,
 	createAPIMeshCredentials,
@@ -46,7 +40,6 @@ jest.mock('../../../helpers', () => ({
 	initSdk: jest.fn().mockResolvedValue({}),
 	initRequestId: jest.fn().mockResolvedValue({}),
 	promptConfirm: jest.fn().mockResolvedValue(true),
-	getFilesInMeshConfig: jest.fn().mockReturnValue([]),
 	importFiles: jest.fn().mockResolvedValue(),
 }));
 jest.mock('../../../lib/devConsole');
@@ -514,7 +507,7 @@ describe('create command tests', () => {
 			files: [
 				{
 					path: './requestParams.json',
-					content: '{"type":"dummyContent"}',
+					content: '{"type":"updatedContent"}',
 				},
 			],
 		};
@@ -530,8 +523,6 @@ describe('create command tests', () => {
 				autoConfirmAction: Promise.resolve(false),
 			},
 		});
-
-		getFilesInMeshConfig.mockReturnValue(['./requestParams.json']);
 
 		importFiles.mockResolvedValueOnce({
 			meshConfig,
@@ -549,7 +540,7 @@ describe('create command tests', () => {
 		    "meshConfig": {
 		      "files": [
 		        {
-		          "content": "{"type":"dummyContent"}",
+		          "content": "{"type":"updatedContent"}",
 		          "path": "./requestParams.json",
 		        },
 		      ],
@@ -602,7 +593,7 @@ describe('create command tests', () => {
 		    "meshConfig": {
 		      "files": [
 		        {
-		          "content": "{"type":"dummyContent"}",
+		          "content": "{"type":"updatedContent"}",
 		          "path": "./requestParams.json",
 		        },
 		      ],
@@ -643,20 +634,14 @@ describe('create command tests', () => {
 			},
 		});
 
-		getFilesInMeshConfig.mockImplementation(() => {
-			throw new Error(
-				'Mesh file names must be less than 25 characters. The following file(s) are invalid: requestJSONParameters.json',
-			);
-		});
-
 		const output = CreateCommand.run();
 
-		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid'));
+		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid.'));
 
 		expect(logSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Mesh file names must be less than 25 characters. The following file(s) are invalid: requestJSONParameters.json",
+		    "Mesh file names must be less than 25 characters. The following file(s) are invalid: requestJSONParameters.json.",
 		  ],
 		]
 	`);
@@ -664,7 +649,7 @@ describe('create command tests', () => {
 		expect(errorLogSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Input mesh config is not valid",
+		    "Input mesh config is not valid.",
 		  ],
 		]
 	`);
@@ -678,26 +663,21 @@ describe('create command tests', () => {
 			},
 		});
 
-		getFilesInMeshConfig.mockImplementation(() => {
-			throw new Error('Please make sure the file names are matching in both places in meshConfig');
-		});
-
 		const output = CreateCommand.run();
 
-		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid'));
+		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid.'));
 
 		expect(logSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Please make sure the file names are matching in both places in meshConfig",
+		    "Please make sure the file names are matching in both places in meshConfig.",
 		  ],
 		]
 	`);
-
 		expect(errorLogSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Input mesh config is not valid",
+		    "Input mesh config is not valid.",
 		  ],
 		]
 	`);
@@ -711,28 +691,21 @@ describe('create command tests', () => {
 			},
 		});
 
-		getFilesInMeshConfig.mockImplementation(() => {
-			throw new Error(
-				'Mesh files must be JavaScript or JSON. Other file types are not supported. The following files are invalid: requestParams.txt',
-			);
-		});
-
 		const output = CreateCommand.run();
 
-		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid'));
+		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid.'));
 
 		expect(logSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Mesh files must be JavaScript or JSON. Other file types are not supported. The following files are invalid: requestParams.txt",
+		    "Mesh files must be JavaScript or JSON. Other file types are not supported. The following file(s) are invalid: requestParams.txt.",
 		  ],
 		]
 	`);
-
 		expect(errorLogSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Input mesh config is not valid",
+		    "Input mesh config is not valid.",
 		  ],
 		]
 	`);
@@ -740,25 +713,19 @@ describe('create command tests', () => {
 
 	test('should fail if the meshConfig and the file are not in the same directory', async () => {
 		parseSpy.mockResolvedValue({
-			args: { file: 'src/commands/__fixtures__/sample_mesh_files.json' },
+			args: { file: 'src/commands/__fixtures__/sample_mesh_invalid_paths.json' },
 			flags: {
 				autoConfirmAction: Promise.resolve(false),
 			},
 		});
 
-		getFilesInMeshConfig.mockImplementation(() => {
-			throw new Error(
-				'Please make sure the files and requestParams.json and sample_mesh_files.json are in the same directory',
-			);
-		});
-
 		const output = CreateCommand.run();
-		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid'));
+		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid.'));
 
 		expect(logSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Please make sure the files and requestParams.json and sample_mesh_files.json are in the same directory",
+		    "Please make sure the files schemaBody.json and sample_mesh_invalid_paths.json are in the same directory.",
 		  ],
 		]
 	`);
@@ -766,7 +733,7 @@ describe('create command tests', () => {
 		expect(errorLogSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Input mesh config is not valid",
+		    "Input mesh config is not valid.",
 		  ],
 		]
 	`);
@@ -779,8 +746,6 @@ describe('create command tests', () => {
 				autoConfirmAction: Promise.resolve(false),
 			},
 		});
-
-		getFilesInMeshConfig.mockReturnValue(['./requestParams.json']);
 
 		importFiles.mockImplementation(() => {
 			throw new Error('Error reading the file');
@@ -847,17 +812,13 @@ describe('create command tests', () => {
 			meshConfig: meshConfig,
 		});
 
+		importFiles.mockResolvedValue(meshConfig);
+
 		parseSpy.mockResolvedValue({
-			args: { file: 'src/commands/__fixtures__/sample_mesh_files.json' },
+			args: { file: 'src/commands/__fixtures__/sample_mesh_with_files_array.json' },
 			flags: {
 				autoConfirmAction: Promise.resolve(false),
 			},
-		});
-
-		getFilesInMeshConfig.mockReturnValue(['./requestParams.json']);
-
-		importFiles.mockResolvedValueOnce({
-			meshConfig,
 		});
 
 		const output = await CreateCommand.run();
@@ -869,33 +830,31 @@ describe('create command tests', () => {
 		  "5678",
 		  "123456789",
 		  {
-		    "meshConfig": {
-		      "files": [
-		        {
-		          "content": "{"type":"dummyContent"}",
-		          "path": "./requestParams.json",
-		        },
-		      ],
-		      "sources": [
-		        {
-		          "handler": {
-		            "JsonSchema": {
-		              "baseUrl": "<json_source__baseurl>",
-		              "operations": [
-		                {
-		                  "field": "<query>",
-		                  "method": "POST",
-		                  "path": "<query_path>",
-		                  "requestSchema": "./requestParams.json",
-		                  "type": "Query",
-		                },
-		              ],
-		            },
+		    "files": [
+		      {
+		        "content": "{"type":"dummyContent"}",
+		        "path": "./requestParams.json",
+		      },
+		    ],
+		    "sources": [
+		      {
+		        "handler": {
+		          "JsonSchema": {
+		            "baseUrl": "<json_source__baseurl>",
+		            "operations": [
+		              {
+		                "field": "<query>",
+		                "method": "POST",
+		                "path": "<query_path>",
+		                "requestSchema": "./requestParams.json",
+		                "type": "Query",
+		              },
+		            ],
 		          },
-		          "name": "<json_source_name>",
 		        },
-		      ],
-		    },
+		        "name": "<json_source_name>",
+		      },
+		    ],
 		  },
 		]
 	`);
@@ -906,7 +865,6 @@ describe('create command tests', () => {
 		  "123456789",
 		]
 	`);
-
 		expect(subscribeCredentialToMeshService.mock.calls[0]).toMatchInlineSnapshot(`
 		[
 		  "1234",
@@ -995,13 +953,11 @@ describe('create command tests', () => {
 		});
 
 		parseSpy.mockResolvedValue({
-			args: { file: 'src/commands/__fixtures__/sample_mesh_files.json' },
+			args: { file: 'src/commands/__fixtures__/sample_mesh_with_files_array.json' },
 			flags: {
 				autoConfirmAction: Promise.resolve(false),
 			},
 		});
-
-		getFilesInMeshConfig.mockReturnValue(['./requestParams.json']);
 
 		importFiles.mockResolvedValueOnce({
 			meshConfig,
