+++ src/commands/api-mesh/__tests__/update.test.js
@@ -17,7 +17,6 @@ jest.mock('../../../helpers', () => ({
 	initSdk: jest.fn().mockResolvedValue({}),
 	initRequestId: jest.fn().mockResolvedValue({}),
 	promptConfirm: jest.fn().mockResolvedValue(true),
-	getFilesInMeshConfig: jest.fn().mockReturnValue([]),
 	importFiles: jest.fn().mockResolvedValue(),
 }));
 jest.mock('@adobe/aio-cli-lib-console', () => ({
@@ -36,13 +35,7 @@ const selectedWorkspace = { id: '123456789', title: 'Workspace01' };
 const { readFile } = require('fs/promises');
 
 const UpdateCommand = require('../update');
-const {
-	initSdk,
-	initRequestId,
-	promptConfirm,
-	getFilesInMeshConfig,
-	importFiles,
-} = require('../../../helpers');
+const { initSdk, initRequestId, promptConfirm, importFiles } = require('../../../helpers');
 const { getMeshId, updateMesh } = require('../../../lib/devConsole');
 
 let logSpy = null;
@@ -82,6 +75,30 @@ describe('update command tests', () => {
 	});
 
 	test('should pass with valid args', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<api_name>',
+						handler: {
+							graphql: {
+								endpoint: '<gql_endpoint>',
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
+		parseSpy.mockResolvedValueOnce({
+			args: { file: 'src/commands/__fixtures__/sample_mesh.json' },
+			flags: {
+				ignoreCache: mockIgnoreCacheFlag,
+				autoConfirmAction: mockAutoApproveAction,
+			},
+		});
+
 		const runResult = await UpdateCommand.run();
 
 		expect(runResult).toMatchInlineSnapshot(`
@@ -117,8 +134,24 @@ describe('update command tests', () => {
 	});
 
 	test('should pass with valid args and ignoreCache flag', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<api_name>',
+						handler: {
+							graphql: {
+								endpoint: '<gql_endpoint>',
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
 		parseSpy.mockResolvedValueOnce({
-			args: { file: 'valid_file_name' },
+			args: { file: 'src/commands/__fixtures__/sample_mesh.json' },
 			flags: {
 				ignoreCache: Promise.resolve(true),
 				autoConfirmAction: mockAutoApproveAction,
@@ -160,8 +193,24 @@ describe('update command tests', () => {
 	});
 
 	test('should pass with valid args if autoConfirmAction flag is set', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<api_name>',
+						handler: {
+							graphql: {
+								endpoint: '<gql_endpoint>',
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
 		parseSpy.mockResolvedValueOnce({
-			args: { file: 'valid_file_name' },
+			args: { file: 'src/commands/__fixtures__/sample_mesh.json' },
 			flags: {
 				ignoreCache: mockIgnoreCacheFlag,
 				autoConfirmAction: Promise.resolve(true),
@@ -204,6 +253,30 @@ describe('update command tests', () => {
 	});
 
 	test('should fail if mesh id is missing', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<api_name>',
+						handler: {
+							graphql: {
+								endpoint: '<gql_endpoint>',
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
+		parseSpy.mockResolvedValueOnce({
+			args: { file: 'src/commands/__fixtures__/sample_mesh.json' },
+			flags: {
+				ignoreCache: mockIgnoreCacheFlag,
+				autoConfirmAction: mockAutoApproveAction,
+			},
+		});
+
 		getMeshId.mockResolvedValue(null);
 		const runResult = UpdateCommand.run();
 
@@ -221,6 +294,30 @@ describe('update command tests', () => {
 	});
 
 	test('should fail if getMeshId api failed', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<api_name>',
+						handler: {
+							graphql: {
+								endpoint: '<gql_endpoint>',
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
+		parseSpy.mockResolvedValueOnce({
+			args: { file: 'src/commands/__fixtures__/sample_mesh.json' },
+			flags: {
+				ignoreCache: mockIgnoreCacheFlag,
+				autoConfirmAction: mockAutoApproveAction,
+			},
+		});
+
 		getMeshId.mockRejectedValue(new Error('getMeshId api failed'));
 		const runResult = UpdateCommand.run();
 
@@ -238,6 +335,30 @@ describe('update command tests', () => {
 	});
 
 	test('should fail if updateMesh method failed', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<api_name>',
+						handler: {
+							graphql: {
+								endpoint: '<gql_endpoint>',
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
+		parseSpy.mockResolvedValueOnce({
+			args: { file: 'src/commands/__fixtures__/sample_mesh.json' },
+			flags: {
+				ignoreCache: mockIgnoreCacheFlag,
+				autoConfirmAction: mockAutoApproveAction,
+			},
+		});
+
 		updateMesh.mockRejectedValueOnce(new Error('dummy_error'));
 
 		const runResult = UpdateCommand.run();
@@ -316,6 +437,30 @@ describe('update command tests', () => {
 	});
 
 	test('should not update if user prompt returns false', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<api_name>',
+						handler: {
+							graphql: {
+								endpoint: '<gql_endpoint>',
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
+		parseSpy.mockResolvedValueOnce({
+			args: { file: 'src/commands/__fixtures__/sample_mesh.json' },
+			flags: {
+				ignoreCache: mockIgnoreCacheFlag,
+				autoConfirmAction: mockAutoApproveAction,
+			},
+		});
+
 		promptConfirm.mockResolvedValueOnce(false);
 
 		const runResult = await UpdateCommand.run();
@@ -332,6 +477,31 @@ describe('update command tests', () => {
 	});
 
 	test('should pass if there are local files in meshConfig i.e., the file is appended in files array', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<json_source_name>',
+						handler: {
+							JsonSchema: {
+								baseUrl: '<json_source__baseurl>',
+								operations: [
+									{
+										type: 'Query',
+										field: '<query>',
+										path: '<query_path>',
+										method: 'POST',
+										requestSchema: './requestParams.json',
+									},
+								],
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
 		let meshConfig = {
 			sources: [
 				{
@@ -355,7 +525,7 @@ describe('update command tests', () => {
 			files: [
 				{
 					path: './requestParams.json',
-					content: '{"type":"dummyContent"}',
+					content: '{"type":"updatedContent"}',
 				},
 			],
 		};
@@ -368,11 +538,9 @@ describe('update command tests', () => {
 		parseSpy.mockResolvedValue({
 			args: { file: 'src/commands/__fixtures__/sample_mesh_files.json' },
 			flags: {
-				autoConfirmAction: Promise.resolve(false),
+				autoConfirmAction: mockAutoApproveAction,
 			},
 		});
-
-		getFilesInMeshConfig.mockReturnValue(['./requestParams.json']);
 
 		importFiles.mockResolvedValueOnce({
 			meshConfig,
@@ -391,7 +559,7 @@ describe('update command tests', () => {
 		    "meshConfig": {
 		      "files": [
 		        {
-		          "content": "{"type":"dummyContent"}",
+		          "content": "{"type":"updatedContent"}",
 		          "path": "./requestParams.json",
 		        },
 		      ],
@@ -418,12 +586,13 @@ describe('update command tests', () => {
 		  },
 		]
 	`);
+
 		expect(output).toMatchInlineSnapshot(`
 		{
 		  "meshConfig": {
 		    "files": [
 		      {
-		        "content": "{"type":"dummyContent"}",
+		        "content": "{"type":"updatedContent"}",
 		        "path": "./requestParams.json",
 		      },
 		    ],
@@ -452,28 +621,48 @@ describe('update command tests', () => {
 	`);
 	});
 
-	test('should fail if the file name is more than 25 characters', async () => {
+	test('should fail if the input mesh config file is inavlid i.e., file name has more than 25 characters', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<json_source_name>',
+						handler: {
+							JsonSchema: {
+								baseUrl: '<json_source__baseurl>',
+								operations: [
+									{
+										type: 'Query',
+										field: '<query>',
+										path: '<query_path>',
+										method: 'POST',
+										requestSchema: './requestJSONParameters.json',
+									},
+								],
+							},
+						},
+					},
+				],
+			},
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
+
 		parseSpy.mockResolvedValue({
 			args: { file: 'src/commands/__fixtures__/sample_mesh_invalid_file_name.json' },
 			flags: {
-				autoConfirmAction: Promise.resolve(false),
+				ignoreCache: mockIgnoreCacheFlag,
+				autoConfirmAction: mockAutoApproveAction,
 			},
-		});
-
-		getFilesInMeshConfig.mockImplementation(() => {
-			throw new Error(
-				'Mesh file names must be less than 25 characters. The following file(s) are invalid: requestJSONParameters.json',
-			);
 		});
 
 		const output = UpdateCommand.run();
 
-		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid'));
+		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid.'));
 
 		expect(logSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Mesh file names must be less than 25 characters. The following file(s) are invalid: requestJSONParameters.json",
+		    "Mesh file names must be less than 25 characters. The following file(s) are invalid: requestJSONParameters.json.",
 		  ],
 		]
 	`);
@@ -481,101 +670,61 @@ describe('update command tests', () => {
 		expect(errorLogSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Input mesh config is not valid",
+		    "Input mesh config is not valid.",
 		  ],
 		]
 	`);
 	});
 
-	test('should fail if the file paths in files array and filenames in sources, transforms, additionalResolvers do not match in mesh config', async () => {
-		parseSpy.mockResolvedValue({
-			args: { file: 'src/commands/__fixtures__/sample_mesh_mismatching_path.json' },
-			flags: {
-				autoConfirmAction: Promise.resolve(false),
+	test('should fail if the import files function fails', async () => {
+		let sampleMesh = {
+			meshConfig: {
+				sources: [
+					{
+						name: '<json_source_name>',
+						handler: {
+							JsonSchema: {
+								baseUrl: '<json_source__baseurl>',
+								operations: [
+									{
+										type: 'Query',
+										field: '<query>',
+										path: '<query_path>',
+										method: 'POST',
+										requestSchema: './requestParams.json',
+									},
+								],
+							},
+						},
+					},
+				],
 			},
-		});
+		};
+		readFile.mockResolvedValue(JSON.stringify(sampleMesh));
 
-		getFilesInMeshConfig.mockImplementation(() => {
-			throw new Error('Please make sure the file names are matching in both places in meshConfig');
-		});
-
-		const output = UpdateCommand.run();
-
-		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid'));
-
-		expect(logSpy.mock.calls).toMatchInlineSnapshot(`
-		[
-		  [
-		    "Please make sure the file names are matching in both places in meshConfig",
-		  ],
-		]
-	`);
-
-		expect(errorLogSpy.mock.calls).toMatchInlineSnapshot(`
-		[
-		  [
-		    "Input mesh config is not valid",
-		  ],
-		]
-	`);
-	});
-
-	test('should fail if the file is of type other than js, json extension', async () => {
-		parseSpy.mockResolvedValue({
-			args: { file: 'src/commands/__fixtures__/sample_mesh_invalid_type.json' },
-			flags: {
-				autoConfirmAction: Promise.resolve(false),
-			},
-		});
-
-		getFilesInMeshConfig.mockImplementation(() => {
-			throw new Error(
-				'Mesh files must be JavaScript or JSON. Other file types are not supported. The following files are invalid: requestParams.txt',
-			);
-		});
-
-		const output = UpdateCommand.run();
-
-		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid'));
-
-		expect(logSpy.mock.calls).toMatchInlineSnapshot(`
-		[
-		  [
-		    "Mesh files must be JavaScript or JSON. Other file types are not supported. The following files are invalid: requestParams.txt",
-		  ],
-		]
-	`);
-
-		expect(errorLogSpy.mock.calls).toMatchInlineSnapshot(`
-		[
-		  [
-		    "Input mesh config is not valid",
-		  ],
-		]
-	`);
-	});
-
-	test('should fail if the meshConfig and the file are not in the same directory', async () => {
 		parseSpy.mockResolvedValue({
 			args: { file: 'src/commands/__fixtures__/sample_mesh_files.json' },
 			flags: {
-				autoConfirmAction: Promise.resolve(false),
+				ignoreCache: mockIgnoreCacheFlag,
+				autoConfirmAction: mockAutoApproveAction,
 			},
 		});
 
-		getFilesInMeshConfig.mockImplementation(() => {
-			throw new Error(
-				'Please make sure the files and requestParams.json and sample_mesh_files.json are in the same directory',
-			);
+		importFiles.mockImplementation(() => {
+			throw new Error('Error reading the file.');
 		});
 
 		const output = UpdateCommand.run();
-		await expect(output).rejects.toEqual(new Error('Input mesh config is not valid'));
+		await expect(output).rejects.toEqual(
+			new Error(
+				'Unable to import the files in the mesh config. Please check the file and try again.',
+			),
+		);
 
 		expect(logSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Please make sure the files and requestParams.json and sample_mesh_files.json are in the same directory",
+		    "Error reading the file.",
 		  ],
 		]
 	`);
@@ -583,7 +732,7 @@ describe('update command tests', () => {
 		expect(errorLogSpy.mock.calls).toMatchInlineSnapshot(`
 		[
 		  [
-		    "Input mesh config is not valid",
+		    "Unable to import the files in the mesh config. Please check the file and try again.",
 		  ],
 		]
 	`);
